make simple syscal system
