+++ .vscode/settings.json
@@ -55,7 +55,8 @@
         "stdint.h": "c",
         "memory_manager.h": "c",
         "ps2_keyboard.h": "c",
-        "idt.h": "c"
+        "idt.h": "c",
+        "string.h": "c"
     },
     "C_Cpp.errorSquiggles": "disabled"
 }
